requirements.txt
tiktoken == 0.12.0
groq == 1.0.0
chromadb ==1.4.1
langchain==1.2.6
fastapi==0.128.0
GitPython == 3.1.46
python-dotenv == 1.2.1
langchain-community==0.4.1
langchain-core == 1.2.7
langchain-huggingface == 1.2.0
uvicorn == 0.40.0
sentence-transformers== 5.2.3

